Fixed concurrency issues.
Database pages were being fetched multiple times. To solve this I had to move to a test-and-set setup while holding the lock. Also moving all logic to the front of get_page() helped avoid fetching data multiple times just for it to be thrown away. Also had to break up NotionPage so it could be initialized with just the page ID.